:recycle: 코드 구조 개편 및 리마인드 이벤트 수신부 TODO로 작성
