--- app/src/main/java/com/umc/ttatta/DebugConfig.kt
@@ -1,9 +1,0 @@
-package com.umc.ttatta
-
-object DebugConfig {
-    const val ID = "app_test"
-    const val PASSWORD = "app_test_pw"
-    const val NAME = "김땡땡"
-    const val NICKNAME = "테스터김"
-    const val EMAIL = "tester@naver.com"
-}
