+++ app/src/main/java/com/umc/ttatta/MainActivity.kt
@@ -1,38 +1,30 @@
 package com.umc.ttatta
 
-import android.app.Activity
-import android.content.ContentValues
 import android.content.Intent
-import android.content.pm.PackageManager
 import android.net.Uri
 import android.os.Build
 import android.os.Bundle
 import android.provider.MediaStore
-import android.view.WindowManager
 import androidx.activity.ComponentActivity
 import androidx.activity.compose.setContent
 import androidx.activity.result.ActivityResultLauncher
 import androidx.activity.result.contract.ActivityResultContracts
 import androidx.activity.viewModels
-import androidx.annotation.RequiresApi
 import androidx.compose.runtime.collectAsState
 import androidx.compose.runtime.getValue
-import androidx.core.view.WindowCompat
-import com.umc.core.repository.UserRepository
+import com.umc.footprint.model.event.RemindEvent
+import com.umc.ttatta.util.createImageUri
+import com.umc.ttatta.util.isCameraPermissionGranted
+import com.umc.ttatta.util.isLocationPermissionGranted
+import com.umc.ttatta.util.isMediaPermissionGranted
+import com.umc.ttatta.util.setStatusBarTransparent
+import com.umc.ttatta.util.uriToFile
 import dagger.hilt.android.AndroidEntryPoint
-import kotlinx.coroutines.CoroutineScope
-import kotlinx.coroutines.Dispatchers
 import kotlinx.coroutines.flow.MutableStateFlow
-import kotlinx.coroutines.launch
 import java.io.File
-import java.io.FileOutputStream
-import java.io.IOException
-import javax.inject.Inject
 
 @AndroidEntryPoint
 class MainActivity : ComponentActivity() {
-    @Inject
-    lateinit var userRepository: UserRepository
     private val viewModel: MainViewModel by viewModels()
 
     private lateinit var imagePickerLauncher: ActivityResultLauncher<Intent>
@@ -44,19 +36,21 @@ class MainActivity : ComponentActivity() {
 
     private var imageUri: Uri? = null
     private val imageFileState = MutableStateFlow<File?>(null)
+    private val remindEventState = MutableStateFlow<RemindEvent?>(null)
 
     override fun onCreate(savedInstanceState: Bundle?) {
         super.onCreate(savedInstanceState)
 
-        // enableDebugMode()
         setLaunchers()
         setStatusBarTransparent()
         setContent {
             val imageFile by imageFileState.collectAsState()
+            val remindEvent by remindEventState.collectAsState()
 
             MainApp(
                 viewModel = viewModel,
                 imageFile = imageFile,
+                remindEvent = remindEvent,
                 onPermissionRequiredInitially = {
                     if (!isLocationPermissionGranted) locationPermissionRequester.launch(
                         android.Manifest.permission.ACCESS_FINE_LOCATION
@@ -78,6 +72,8 @@ class MainActivity : ComponentActivity() {
                 },
             )
         }
+
+        resolveIntent(intent = intent)
     }
 
     private fun setLaunchers() {
@@ -126,66 +122,7 @@ class MainActivity : ComponentActivity() {
         }
     }
 
-    private fun enableDebugMode() {
-        CoroutineScope(Dispatchers.IO).launch {
-            if (!userRepository.isIdAlreadyOccupied(DebugConfig.ID)) {
-                userRepository.join(
-                    id = DebugConfig.ID,
-                    password = DebugConfig.PASSWORD,
-                    name = DebugConfig.NAME,
-                    nickname = DebugConfig.NICKNAME,
-                    email = DebugConfig.EMAIL
-                )
-            }
-
-            userRepository.login(
-                id = DebugConfig.ID,
-                password = DebugConfig.PASSWORD
-            )
-
-            viewModel.checkLogin()
-        }
+    private fun resolveIntent(intent: Intent) {
+        // TODO: 여기에 알림 인텐트 처리 로직 추가
     }
 }
-
-private fun ComponentActivity.setStatusBarTransparent() {
-    window.apply {
-        WindowCompat.setDecorFitsSystemWindows(this, false)
-        setFlags(
-            WindowManager.LayoutParams.FLAG_LAYOUT_NO_LIMITS,
-            WindowManager.LayoutParams.FLAG_LAYOUT_NO_LIMITS
-        )
-    }
-}
-
-private fun Activity.createImageUri(): Uri? {
-    val contentValues = ContentValues().apply {
-        put(MediaStore.Images.Media.DISPLAY_NAME, "photo_${System.currentTimeMillis()}.jpg")
-        put(MediaStore.Images.Media.MIME_TYPE, "image/jpeg")
-    }
-    return contentResolver.insert(MediaStore.Images.Media.EXTERNAL_CONTENT_URI, contentValues)
-}
-
-private fun Activity.uriToFile(uri: Uri): File? {
-    val file = File(cacheDir, "temp_image_${System.currentTimeMillis()}.jpg") // 내부 캐시 디렉토리에 저장
-    try {
-        contentResolver.openInputStream(uri)?.use { inputStream ->
-            FileOutputStream(file).use { outputStream ->
-                inputStream.copyTo(outputStream) // 스트림을 복사
-            }
-        }
-        return file
-    } catch (e: IOException) {
-        e.printStackTrace()
-    }
-    return null
-}
-
-private val Activity.isCameraPermissionGranted get() =
-    checkSelfPermission(android.Manifest.permission.CAMERA) == PackageManager.PERMISSION_GRANTED
-
-private val Activity.isLocationPermissionGranted get() =
-    checkSelfPermission(android.Manifest.permission.ACCESS_FINE_LOCATION) == PackageManager.PERMISSION_GRANTED
-
-private val Activity.isMediaPermissionGranted @RequiresApi(Build.VERSION_CODES.Q) get() =
-    checkSelfPermission(android.Manifest.permission.ACCESS_MEDIA_LOCATION) == PackageManager.PERMISSION_GRANTED
