+++ app/src/main/java/com/umc/ttatta/model/CategoryRoutingInfo.kt
@@ -1,0 +1,5 @@
+package com.umc.ttatta.model
+
+class CategoryRoutingInfo(
+    val showTopBar: Boolean,
+)
