+++ app/src/main/java/com/umc/ttatta/model/ChallengeRoutingInfo.kt
@@ -1,0 +1,6 @@
+package com.umc.ttatta.model
+
+class ChallengeRoutingInfo(
+    val isPointGranted: Boolean,
+    val isPoppedFromRecord: Boolean,
+)
