+++ app/src/main/java/com/umc/ttatta/model/RecordEntryInfo.kt
@@ -1,0 +1,6 @@
+package com.umc.ttatta.model
+
+class RecordEntryInfo(
+    val mode: RecordRoutingOption,
+    val challengeId: Long?,
+)
