+++ app/src/main/java/com/umc/ttatta/model/RecordRoutingInfo.kt
@@ -1,0 +1,8 @@
+package com.umc.ttatta.model
+
+import java.io.File
+
+class RecordRoutingInfo(
+    val image: File,
+    val challengeId: Long?,
+)
