+++ app/src/main/java/com/umc/ttatta/model/RecordRoutingOption.kt
@@ -1,0 +1,6 @@
+package com.umc.ttatta.model
+
+enum class RecordRoutingOption {
+    CAMERA,
+    GALLERY,
+}
